Refactored API resources.
